requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_py-xgboost-mutex=2.0=cpu_0
_pytorch_select=0.2=gpu_0
alembic=1.7.5=pypi_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=20.1.0=py37h27cfd23_1
async_generator=1.10=py_0
attrs=21.2.0=pyhd3eb1b0_0
autopage=0.4.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
boost=1.70.0=py37h9de70de_1
boost-cpp=1.70.0=ha2d47e9_0
bottleneck=1.3.2=py37heb32a55_1
bzip2=1.0.6=3
ca-certificates=2021.10.26=h06a4308_2
cairo=1.16.0=h18b612c_1001
certifi=2021.10.8=py37h06a4308_0
cffi=1.14.6=py37h400218f_0
cliff=3.10.0=pypi_0
cmaes=0.8.2=pypi_0
cmd2=2.3.2=pypi_0
conda-pack=0.7.1=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.4=h0708190_1
gensim=4.1.2=pypi_0
gettext=0.19.8.1=h0b5b191_1005
glib=2.68.4=h9c3ff4c_0
glib-tools=2.68.4=h9c3ff4c_0
greenlet=1.1.2=pypi_0
icu=58.2=hf484d3e_1000
importlib-metadata=4.8.1=py37h06a4308_0
importlib-resources=5.4.0=pypi_0
importlib_metadata=4.8.1=hd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
ipykernel=6.4.1=py37h06a4308_1
ipython=7.29.0=py37hb070fc8_0
ipython_genutils=0.2.0=py_1
jbig=2.1=0
jedi=0.18.0=py37h06a4308_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h36c2ea0_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.9.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=1_h86c2bf4_netlib
libcblas=3.9.0=5_h92ddd45_netlib
libdeflate=1.7=h7f98852_5
libffi=3.3=h58526e2_2
libgcc-ng=11.1.0=hc902ee8_8
libgfortran-ng=11.1.0=h69a702a_8
libgfortran5=11.1.0=h6c583b3_8
libglib=2.68.4=h3e27bee_0
libgomp=11.1.0=hc902ee8_8
libiconv=1.16=h516909a_0
liblapack=3.9.0=5_h92ddd45_netlib
libopenblas=0.3.17=pthreads_h8fe5266_1
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.1.0=h56837e0_8
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.1=h7f98852_0
libxcb=1.12=1
libxgboost=1.5.0=h9c3ff4c_1
libxml2=2.9.9=h13577e0_2
lz4-c=1.9.3=h9c3ff4c_1
mako=1.1.6=pypi_0
markupsafe=2.0.1=py37h27cfd23_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.3.0=py37h54f3939_0
mkl_random=1.1.1=py37h0573a6f_0
nbclient=0.5.8=pyhd8ed1ab_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd8ed1ab_0
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
ninja=1.10.2=hff7bd54_1
notebook=6.4.6=py37h06a4308_0
numexpr=2.7.3=py37hb2eb853_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=pyhd3eb1b0_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f8727e_0
optuna=2.10.0=pypi_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.4=py37h8c16a72_0
pandoc=2.16.1=h7f98852_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.7.1=pyh9f0ad1d_0
pbr=5.8.0=pypi_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.3.1=py37h2c7a002_0
pip=23.3=pypi_0
pixman=0.38.0=h516909a_1003
prettytable=2.4.0=pypi_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.22=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pycairo=1.19.1=py37h708ec4a_0
pycparser=2.20=py_2
pygments=2.10.0=pyhd8ed1ab_0
pyparsing=3.0.5=pyhd8ed1ab_0
pyperclip=1.8.2=pypi_0
pyrsistent=0.18.0=py37heee7806_0
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.2.0=cuda100py37h938c94c_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=22.2.1=py37h295c915_1
rdkit=2019.03.3=py37hb31dc5d_0
readline=8.1=h46c0cb4_0
scikit-learn=1.0.1=py37h51133e4_0
scipy=1.7.3=py37hf2a6cf1_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_0
smart-open=5.2.1=pypi_0
sqlalchemy=1.4.27=pypi_0
sqlite=3.36.0=h9cd32fc_0
stevedore=3.5.0=pypi_0
terminado=0.9.4=py37h06a4308_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h21135ba_0
tornado=6.1=py37h27cfd23_0
traitlets=4.3.3=pyhd8ed1ab_2
typing_extensions=3.10.0.2=pyh06a4308_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.37.0=pyhd3eb1b0_1
xgboost=1.6.2=pypi_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.6.0=pyhd3eb1b0_0
zlib=1.2.11=0
zstd=1.5.0=ha95c52a_0
